// types/service.ts
export interface NoticeInterface {
  id: number;
  created: string;
  title: string;
  description: string;
}

export interface FaqInterface {
  id: number;
  title: string;
  description: string;
  created: string;
}

export interface MainInfoInterface {
  id: number;
  title: string;
  description: string;
  created: string;
}